# adapter/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Service API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      responses:
        '200':
          description: Service health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /payments:
    post:
      operationId: createPayment
      summary: Create payment
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XAPIKey'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '200':
          description: Existing idempotent payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /payments/{payment_id}:
    get:
      operationId: getPayment
      summary: Get payment
      parameters:
        - name: payment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          $ref: '#/components/responses/Error'
  /provider-webhooks/{provider_name}:
    post:
      operationId: providerWebhook
      summary: Provider webhook
      parameters:
        - name: provider_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '400':
          $ref: '#/components/responses/Error'
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
    XAPIKey:
      name: X-API-Key
      in: header
      required: true
      schema:
        type: string
    XTimestamp:
      name: X-Timestamp
      in: header
      required: true
      schema:
        type: string
    XSignature:
      name: X-Signature
      in: header
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok
    CreatePaymentRequest:
      type: object
      required: [order_id, amount_minor, currency]
      properties:
        order_id:
          type: string
        amount_minor:
          type: integer
          format: int64
          minimum: 1
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: RUB
        description:
          type: string
        payment_method:
          type: string
          example: bank_card
        customer:
          $ref: '#/components/schemas/Customer'
    Customer:
      type: object
      properties:
        email:
          type: string
          format: email
    PaymentResponse:
      type: object
      required: [payment_id, order_id, amount_minor, currency, status]
      properties:
        payment_id:
          type: string
          format: uuid
        order_id:
          type: string
        amount_minor:
          type: integer
          format: int64
        currency:
          type: string
        status:
          $ref: '#/components/schemas/PaymentStatus'
        payment_url:
          type: string
    PaymentStatus:
      type: string
      enum: [pending, succeeded, failed, canceled, refunded]
    WebhookResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string